[feat][coordinator] Runtime control for balance region knobs (limit_score_diff, default_store_region_size)
- balance_region_limit_score_diff: register a strictly-positive gflags
  validator so it can be changed at runtime via brpc /flags.
- balance_region_default_store_region_size: new int64 gflag that overrides
  the config file value when > 0 (0 keeps the config-file semantics);
  ConfigHelper getter widened to int64 so overrides beyond 2GB work.
- Wire both knobs into the coordinator ControlConfig whitelist via new
  Helper::HandleInt64/HandleDoubleControlConfigVariable, mirroring the bool
  handler contract (query mode, invalid value sets is_error_occurred and
  keeps the flag, identical value sets is_already_set).
- Unit tests cover getter precedence (flag > config > built-in 256MB),
  the two handler contracts, and validator rejection of nonsense values.

Neither value is persisted: a coordinator restart falls back to the config
file / built-in default, matching every existing ControlConfig variable.

+++ src/config/config_helper.h
@@ -48,7 +48,7 @@ class ConfigHelper {
   static std::string GetBalanceRegionInspectionTimePeriod();
   static float GetBalanceRegionCountRatio();
   static int32_t GetBalanceRegionDefaultIndexRegionSize();
-  static int32_t GetBalanceRegionDefaultStoreRegionSize();
+  static int64_t GetBalanceRegionDefaultStoreRegionSize();
 
   static int32_t GetWorkerThreadNum();
   static double GetWorkerThreadRatio();
